See pipeline.md for a description of all changes. The main changes are: - Curved-sky Born convergence projector directly on HEALPix pixels   (cmb_lensing.py), replacing the flat-sky tangent-plane approximation. - PNG implementation (f_NL) with or without universality relation. - Full oversampling pipeline (init/evol/ptcl/paint grids, model.py,   bricks.py) with PNG anti-aliasing (add_png re-band-limited to the   init Nyquist) - Finger-of-God higher-derivative LOS bias bnpar implemented as a   velocity term (lagrangian_fog_velocity), added before RSD projection. - Free per-radial-bin mean density (ngbars) to marginalise the   integral constraint on the lightcone. - Background emulator (nbody.py): bilinear (Omega_m, a) interpolation   replacing in-graph jax_cosmo ODE solves that exhausted LLVM memory   on large MCMC runs. - Exact cut-sky pixel-space likelihood via signal-eigenmode (KL)   compression (pixel_exact mode), alongside the existing diagonal   MASTER-decoupled mode. - Free shot-noise amplitude s_e (optional, off by default) and   completeness-cut survey mask/radial selection from AbacusLensing   randoms.

+++ src/desi_cmb_fli/chains.py
@@ -380,7 +380,7 @@ class Chains(Samples):
         if thinning is None:
             n_split = 1
         else:
-            n_split = max(np.rint(length / thinning), 1)
+            n_split = max(int(np.rint(length / thinning)), 1)
 
         if moment is None:
